change type of M to float

+++ include/Blit.h
@@ -6,7 +6,7 @@
 
 #include <stdint.h>
 
-void msinc(float * out, float * lfSin, float * lfCos, float * hfSin, float * hfCos, int16_t M);
-int16_t blit_m(float f);
+void msinc(float * out, float * lfSin, float * lfCos, float * hfSin, float * hfCos, float M);
+float blit_m(float f);
 
 #endif // BLIT_H_
